Moved this back. At the moment proxies need too much stuff from AICentral. So pointless having this in core
